% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestors.R
\name{.ancestors}
\alias{.ancestors}
\title{Get parent nodes of a given node}
\usage{
.ancestors(node, onto, include_self = TRUE)
}
\arguments{
\item{node}{the name of a node on the ontology}

\item{onto}{the considered section of the cell ontology as an igraph object}

\item{include_self}{if \code{TRUE}, then the function returns also \code{node}}
}
\value{
returns a vector with the names of the parents
}
\description{
Get parent nodes of a given node
}
\author{
Daniela Corbetta
}
