% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{.children}
\alias{.children}
\title{Get children nodes of a given node}
\usage{
.children(node, onto, leaf = TRUE)
}
\arguments{
\item{node}{the name of a node on the ontology}

\item{onto}{the considered section of the cell ontology as an igraph object}

\item{leaf}{if \code{TRUE}, then the function returns only the leaf nodes that are children of the given one.
Otherwise it returns all the children.}
}
\value{
returns a vector with the names of the children.
}
\description{
Get children nodes of a given node
}
\author{
Daniela Corbetta
}
