% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSets.R
\name{.predSets}
\alias{.predSets}
\title{Get hierarchical prediction sets with given lambda for a single line}
\usage{
.predSets(lambda, pred, onto)
}
\arguments{
\item{lambda}{score/probability treshold}

\item{pred}{a named vectors with the estimated probabilities for each cell type. Names
have to correspond to names of the leaf nodes in the ontology}

\item{onto}{the considered section of the cell ontology as an igraph object}
}
\value{
vector with names of the selected leaf nodes
}
\description{
Get hierarchical prediction sets with given lambda for a single line
}
\author{
Daniela Corbetta
}
