% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{.scores}
\alias{.scores}
\title{Get scores associated with a node}
\usage{
.scores(pred, int_node, onto)
}
\arguments{
\item{pred}{a named vectors with the estimated probabilities for each cell type. Names
have to correspond to names of the leaf nodes in the ontology}

\item{int_node}{the name of a node on the ontology}

\item{onto}{the considered section of the cell ontology as an igraph object}
}
\value{
returns the score
}
\description{
The score is given by the sum of the estimated probabilities of the leaf nodes that
are children of the given node
}
\author{
Daniela Corbetta
}
