% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConformalPredSets.R
\name{getConformalPredSets}
\alias{getConformalPredSets}
\title{Get prediction sets with split conformal inference}
\usage{
getConformalPredSets(p.cal, p.test, y.cal, alpha)
}
\arguments{
\item{p.cal}{a nxK matrix of category scores for the cells in the calibration data.
Needs to have colnames for each category}

\item{p.test}{a mxK matrix of category scores for the cells in the test data}

\item{y.cal}{a vector of length n with the labels of the cells in the calibration data}

\item{alpha}{a number between 0 and 1 that indicates the allowed miscoverage}
}
\value{
The function \code{getConformalPredSets} returns a list of length equal to the number of cells in the test data.
Each element of the list contains the prediction set for that cell.
}
\description{
Let K be the total number of distinct cell type labels and n, m
the number of cells in the calibration and in the test data, respectively.
This function takes as input two matrices: a matrix \code{n x K} and
a matrix \code{m x K} with the estimated
probabilities for each cell in the calibration and in the test data, respectively.
It returns a list with the prediction sets for each cell in the test data.
}
\references{
For reference on split conformal prediction, refer to section 1 of
Angelopoulos, Anastasios N., and Stephen Bates. "A gentle introduction to conformal prediction and distribution-free uncertainty quantification." arXiv preprint arXiv:2107.07511 (2021).
}
\author{
Daniela Corbetta
}
