% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHierarchicalPredSets.R
\name{getHierarchicalPredSets}
\alias{getHierarchicalPredSets}
\title{Get hierarchical prediction sets exploiting the cell ontology}
\usage{
getHierarchicalPredSets(p.cal, p.test, y.cal, onto, alpha, lambdas)
}
\arguments{
\item{p.cal}{a nxK matrix of category scores for the cells in the calibration data.
Needs to have colnames for each category}

\item{p.test}{a mxK matrix of category scores for the cells in the test data}

\item{y.cal}{a vector of length n with the labels of the cells in the calibration data}

\item{onto}{the considered section of the cell ontology}

\item{alpha}{a number between 0 and 1 that indicates the allowed miscoverage}

\item{lambdas}{a vector of possible lambda values to be considered}
}
\value{
The function \code{getHierarchicalPredSets} returns a list of length equal to the number of cells in the test data.
Each element of the list contains the prediction set for that cell.
}
\description{
Let K be the total number of distinct cell type labels and n, m
the number of cells in the calibration and in the test data, respectively.
This function takes as input two matrices: a matrix \code{n x K} and
a matrix \code{m x K} with the estimated
probabilities for each cell in the calibration and in the test data, respectively.
It returns a list with the prediction sets for each cell in the test data.
}
\references{
For reference on conformal risk control, see
Angelopoulos, Anastasios N., et al. "Conformal risk control." arXiv preprint arXiv:2208.02814 (2022).
}
\author{
Daniela Corbetta
}
