% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnCommonAncestor.R
\name{returnCommonAncestor}
\alias{returnCommonAncestor}
\title{Return the common ancestor of the labels in the prediction set}
\usage{
returnCommonAncestor(pred.set, onto)
}
\arguments{
\item{pred.set}{character vector containing the labels in the prediction set}

\item{onto}{ontology as an igraph object}
}
\value{
the common ancestor of the labels in \code{pred.set}, according to
the ontology \code{onto}
}
\description{
This function takes as input a prediction set and an
ontology and returns the common ancestor of the labels in the prediction set.
It is useful when using hierarchical
}
\examples{
library(igraph)
# Let's build a random ontology
onto <- graph_from_literal(animal-+dog:cat, cat-+british:persian,
dog-+cocker:retriever, retriever-+golden:labrador)
# Let's consider this prediction set
pred.set <- c("golden", "labrador", "cocker")
com.anc <- returnCommonAncestor(pred.set, onto)
}
