% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonAncestor.R
\name{getCommonAncestor}
\alias{getCommonAncestor}
\title{Return the common ancestor of the labels in the prediction set}
\usage{
getCommonAncestor(pred_set, onto)
}
\arguments{
\item{pred_set}{character vector containing the labels in the prediction set}

\item{onto}{ontology as an igraph object}
}
\value{
the common ancestor of the labels in \code{pred_set}, according to
the ontology \code{onto}
}
\description{
This function takes as input a prediction set and an
ontology and returns the common ancestor of the labels in the prediction set.
It is useful when using hierarchical
}
\examples{
library(igraph)
# Let's build a random ontology
onto <- graph_from_literal(
    animal-+dog:cat, cat-+british:persian,
    dog-+cocker:retriever, retriever-+golden:labrador
)
# Let's consider this prediction set
pred_set <- c("golden", "labrador", "cocker")
com_anc <- getCommonAncestor(pred_set, onto)
}
