% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResult.R
\name{plotResult}
\alias{plotResult}
\title{Plot prediction sets}
\usage{
plotResult(
  pred_set,
  onto,
  onto_cache = NULL,
  probs = NULL,
  col_grad = c("lemonchiffon", "orange", "darkred"),
  attrs = NULL,
  k = 4,
  title = NULL,
  add_scores = TRUE,
  ...
)
}
\arguments{
\item{pred_set}{character vector containing the labels in the prediction set}

\item{onto}{an \code{igraph} object representing the ontology}

\item{onto_cache}{optional precomputed cache from \code{precomputeOnto()}. If \code{NULL},
the cache is computed automatically (and discarded after the call).}

\item{probs}{numeric vector of estimated probabilities for the classes. The
names of \code{probs} should correspond to node names in \code{onto}}

\item{col_grad}{character vector of colors used to highlight the classes. If
\code{probs} is provided, this should contain a color gradient; otherwise,
a single color can be supplied}

\item{attrs}{attrs list of additional graphical attributes passed to
\code{plot()}}

\item{k}{integer number of decimal digits to consider in \code{probs}}

\item{title}{title of the plot}

\item{add_scores}{Logical. If \code{TRUE}, estimated probabilities are
added to the name of the classes}

\item{...}{additional graphical parameters passed to \code{plot()}}
}
\value{
A plot of the ontology with the labels in the prediction set
highlighted.
}
\description{
This function takes as input a prediction set and an
ontology and plots the ontology, highlighting the labels included in the set.
}
\examples{
library(igraph)
# Let's build a random ontology
onto <- graph_from_literal(
    animal -+dog:cat, cat -+british:persian,
    dog -+cocker:retriever, retriever -+golden:labrador
)
# Let's consider this prediction set
pred_set <- c("golden", "labrador", "cocker")
plotResult(pred_set, onto,
    col_grad = "pink", add_scores = FALSE,
    title = "Prediction set"
)

}
