% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResult.R
\name{plotResult}
\alias{plotResult}
\title{Plot prediction sets}
\usage{
plotResult(
  pred_set,
  onto,
  probs = NULL,
  col_grad = c("lemonchiffon", "orange", "darkred"),
  attrs = NULL,
  k = 4,
  title = NULL,
  add_scores = TRUE,
  ...
)
}
\arguments{
\item{pred_set}{character vector containing the labels in the prediction set}

\item{onto}{ontology as an igraph}

\item{probs}{estimated probabilities for the classes}

\item{col_grad}{color to use to highlight the classes}

\item{attrs}{other grtaphical attributes}

\item{k}{number of decimal digits to consider in \code{probs}}

\item{title}{title of the plot}

\item{add_scores}{boolean. If \code{TRUE}, estimated probabilities are
added to the name of the classes}

\item{...}{general commands to be sent to plot}
}
\value{
a plot of the ontology with the considered classes colored
}
\description{
This function takes as input a prediction set and an
ontology and plots the ontology, highlighting the labels included in the set.
}
\examples{
library(igraph)
# Let's build a random ontology
onto <- graph_from_literal(
    animal -+dog:cat, cat -+british:persian,
    dog -+cocker:retriever, retriever -+golden:labrador
)
# Let's consider this prediction set
pred_set <- c("golden", "labrador", "cocker")
plotResult(pred_set, onto,
    col_grad = "pink", add_scores = FALSE,
    title = "Prediction set"
)

}
