% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResult.R
\name{plotResult}
\alias{plotResult}
\title{Plot prediction sets}
\usage{
plotResult(
  pred.set,
  onto,
  probs = NULL,
  col.grad = c("lemonchiffon", "orange", "darkred"),
  attrs = NULL,
  k = 4,
  title = NULL,
  add.scores = TRUE,
  ...
)
}
\arguments{
\item{pred.set}{character vector containing the labels in the prediction set}

\item{onto}{ontology as an igraph}

\item{probs}{estimated probabilities for the classes}

\item{col.grad}{color to use to highlight the classes}

\item{attrs}{other grtaphical attributes}

\item{k}{number of decimal digits to consider in \code{probs}}

\item{title}{title of the plot}

\item{add.scores}{boolean. If \code{TRUE}, estimated probabilities are
added to the name of the classes}
}
\value{
a plot of the ontology with the considered classes colored
}
\description{
This function takes as input a prediction set and an
ontology and plots the ontology, highlighting the labels included in the set.
}
